+++ frontend/components/CheckBoxes/connector.js
@@ -1,5 +1,18 @@
 import { connect } from 'react-redux';
+import { getCartItems } from '@shopgate/pwa-common-commerce/cart/selectors';
+import { termsToDisplay } from '../../selectors';
 import { updateCheckoutIsOrderable } from '../../action-creators';
+
+/**
+ *
+ * @param {Object} state The current state.
+ * @param {Object} props The props.
+ * @returns {Object}
+ */
+const mapStateToProps = (state, props) => ({
+  productSpecificTerms: termsToDisplay(state),
+  cartItems: getCartItems(state, props),
+});
 
 /**
  * Connects the dispatch function to a callable funcion in props.
@@ -10,4 +23,4 @@ const mapDispatchToProps = dispatch => ({
   setCheckoutIsOrderable: isOrderable => dispatch(updateCheckoutIsOrderable(isOrderable)),
 });
 
-export default connect(null, mapDispatchToProps);
+export default connect(mapStateToProps, mapDispatchToProps);