+++ frontend/portals/CartItemListAfter/index.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import CheckBoxes from '../../components/CheckBoxes';
+import Checkboxes from '../../components/Checkboxes';
 import getConfig from '../../helpers/getConfig';
 
 export default () => {
   const config = getConfig();
 
-  if (config.checkBoxValues.length === 0) {
+  if (config.checkboxValues.length === 0 && config.productSpecificCheckboxValues.length === 0) {
     return null;
   }
-  return (<CheckBoxes config={config} />);
+  return (<Checkboxes config={config} />);
 };