+++ frontend/selectors/index.js
@@ -1,4 +1,6 @@
 import { createSelector } from 'reselect';
+import { getCartProducts } from '@shopgate/pwa-common-commerce/cart/selectors';
+import getConfig from '../helpers/getConfig';
 
 export const getTermsStatus = createSelector(
   state => state,
@@ -10,3 +12,19 @@ export const getTermsStatus = createSelector(
     return extensionState.isOrderable;
   }
 );
+
+export const termsToDisplay = createSelector(
+  getCartProducts,
+  (cartProducts) => {
+    if (!cartProducts) {
+      return null;
+    }
+    const { productSpecificCheckboxValues } = getConfig();
+    const productIds = [];
+    cartProducts.map(cartProduct => productIds.push(cartProduct.product.id));
+    const parsedTerms = productSpecificCheckboxValues.filter((value =>
+      productIds.includes(value.productId)));
+    return parsedTerms;
+  }
+);
+